Default HostIntercept on PlexIcon on iMovie
1. First, if you use an Ethernet connection with your AppleTV,
disconnect it from the other end (not your ATV end).

2: Next, connect (the/an) Ethernet cable to somewhere to go nowhere
(e.g. a laptop or desktop that is powered on but won't allow the
AppleTV to connect to the internet!). If you use a wireless connection
you should also do this, or you can disconnect your router from the
outside world (unplug your DSL cable or equivalent) but keep it powered
on so it hands out IP addresses!

3: Next, reset your AppleTV to factory settings (Settings > General >
Reset > Reset All Settings).

4. When your AppleTV has finished, follow through the setup, pressing
MENU when you reach the wireless setup screen so you don't connect to
the web. Afterwards you should only see the Computers and Settings
options, since it hasn't connected to the web and doesn't have any
additional options.

5. Now go to Settings > General > Network and set your DNS settings as
per the PlexConnect installation guide. Once that's done reconnect your
router / ethernet lead and let your AppleTV loose on the web! It should
connect to Plex as well as modifying the Trailers icon to a Plex icon!